cramjam-python/benchmark-requirements.txt
pytest-benchmark==4.0.0
python-snappy==0.5.4
lz4==3.1.0
brotlipy==0.7.0
zstd==1.5.4.0
isal==0.11.1
numpy
memory-profiler
